+++ be/src/runtime/runtime_state.h
@@ -178,6 +178,10 @@ public:
                _query_options.check_overflow_for_decimal;
     }
 
+    bool enable_strict_mode() const {
+        return _query_options.__isset.enable_strict_cast && _query_options.enable_strict_cast;
+    }
+
     bool enable_decimal256() const {
         return _query_options.__isset.enable_decimal256 && _query_options.enable_decimal256;
     }
