+++ be/src/udf/udf.cpp
@@ -60,6 +60,7 @@ std::unique_ptr<FunctionContext> FunctionContext::clone() {
     new_context->_constant_cols = _constant_cols;
     new_context->_fragment_local_fn_state = _fragment_local_fn_state;
     new_context->_check_overflow_for_decimal = _check_overflow_for_decimal;
+    new_context->_enable_strict_mode = _enable_strict_mode;
     new_context->_string_as_jsonb_string = _string_as_jsonb_string;
     new_context->_jsonb_string_as_string = _jsonb_string_as_string;
     return new_context;
