+++ be/src/vec/data_types/serde/data_type_serde.h
@@ -261,6 +261,34 @@ public:
 public:
     DataTypeSerDe(int nesting_level = 1) : _nesting_level(nesting_level) {};
     virtual ~DataTypeSerDe();
+
+    // All types can override this function
+    // When this function is called, column should be of the corresponding type
+    virtual Status from_string(StringRef& str, IColumn& column,
+                               const FormatOptions& options) const {
+        return Status::NotSupported("from_string is not supported");
+    }
+
+    // Similar to from_string, but in strict mode, we should not handle errors.
+    // If conversion from string fails, we should return immediately
+    virtual Status from_string_strict_mode(StringRef& str, IColumn& column,
+                                           const FormatOptions& options) const {
+        return from_string(str, column, options);
+    }
+
+    // Override functions with _batch suffix only when performance is critical
+    // Only used for basic data types, such as Ip, Date, Number, etc.
+    virtual Status from_string_batch(const ColumnString& str, ColumnNullable& column,
+                                     const FormatOptions& options) const {
+        return Status::NotSupported("from_string is not supported");
+    }
+
+    // For strict mode, we should not have nullable columns, as we will directly report errors when string conversion fails instead of handling them
+    virtual Status from_string_strict_mode_batch(const ColumnString& str, ColumnNullable& column,
+                                                 const FormatOptions& options) const {
+        return Status::NotSupported("from_string is not supported");
+    }
+
     // Text serializer and deserializer with formatOptions to handle different text format
     virtual Status serialize_one_cell_to_json(const IColumn& column, int64_t row_num,
                                               BufferWritable& bw, FormatOptions& options) const = 0;
