+++ be/src/vec/exprs/vexpr_context.cpp
@@ -130,6 +130,7 @@ int VExprContext::register_function_context(RuntimeState* state, const DataTypeP
                                             const std::vector<DataTypePtr>& arg_types) {
     _fn_contexts.push_back(FunctionContext::create_context(state, return_type, arg_types));
     _fn_contexts.back()->set_check_overflow_for_decimal(state->check_overflow_for_decimal());
+    _fn_contexts.back()->set_enable_strict_mode(state->enable_strict_mode());
     return static_cast<int>(_fn_contexts.size()) - 1;
 }
